requirements.txt
fastapi == 0.135.1
uvicorn == 0.42.0
streamlit == 1.55.0
requests == 2.32.5
langchain == 0.2.14
langchain-community == 0.2.12
langchain-core == 0.2.43
langchain-huggingface == 0.0.3
langchain-text-splitters == 0.2.2
transformers == 4.40.0
torch == 2.10.0
accelerate == 1.13.0
sentence-transformers == 2.6.1
faiss-cpu == 1.13.2
pypdf == 6.9.1
python-dotenv == 1.2.2
pydantic == 2.12.5